move command_line_interface.rb to bin to lib
This will support other programs running other business logic before
starting the que locker and workers.

```
exit_code =
  Que::CommandLineInterface.parse(
    args: ARGV.dup,
    output: $stdout,
  )

exit(exit_code)
```
