+++ lib/que/command_line_interface.rb
