+++ spec/que/command_line_interface_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require_relative 'command_line_interface'
 
 require 'digest/md5'
+require 'que/command_line_interface'
 
 describe Que::CommandLineInterface do
   LOADED_FILES = {}
